+++ lib/ui/screens/auth/auth_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '../../../features/sign_in/presentations/view/sign_in_screen.dart';
 import '../../shared/extensions/build_context.dart';
-import '../sign_in/sign_in_screen.dart';
 import '../sign_up/sign_up_screen.dart';
 
 class AuthScreen extends StatelessWidget {
